requirements.txt
Django==1.7
docker-py
